requirements.txt
annotated-types==0.7.0
anyio==4.9.0
blinker==1.9.0
certifi==2025.6.15
click==8.2.1
colorama==0.4.6
distlib==0.3.9
filelock==3.18.0
Flask==3.1.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
packaging==25.0
Pillow==11.2.1
platformdirs==4.3.8
pydantic==2.11.7
pydantic_core==2.33.2
python-dotenv==1.1.1
replicate==1.0.7
setuptools==80.9.0
sniffio==1.3.1
typing-inspection==0.4.1
typing_extensions==4.14.1
virtualenv==20.31.2
Werkzeug==3.1.3
transformers==4.40.1
torch==2.3.0
deep-translator==1.11.4
opencv-python==4.12.0.88
mediapipe==0.10.14
numpy==2.2.0
